Abstract 'submitScore' function
