+++ src/pages/Judging/View.js
@@ -4,7 +4,7 @@ import HeroPage, { Loading, JudgingNotOpen } from '../../components/HeroPage'
 import ViewProject from '../../components/Judging/ViewProject'
 import { A } from '../../components/Typography'
 import ErrorBanner from '../../components/ErrorBanner'
-import { getLivesiteDoc, projectsRef, db, applicantsRef } from '../../utility/firebase'
+import { getLivesiteDoc, submitGrade, projectsRef, db, applicantsRef } from '../../utility/firebase'
 import { useAuth } from '../../utility/Auth'
 import { defaultScoreFromRubric, isUngraded } from '../../utility/Constants'
 
@@ -51,23 +51,7 @@ export default ({ id }) => {
     } else if (!isSubmitting) {
       setFormError(false)
       setIsSubmitting(true)
-      try {
-        await db.runTransaction(async transaction => {
-          const projectDoc = await transaction.get(projectsRef.doc(id))
-          if (!projectDoc.exists) {
-            setIsSubmitting(false)
-            console.err('Project does not exist')
-            setShowError(true)
-            return
-          }
-          const oldGrades = projectDoc.data().grades
-          const grades = { ...oldGrades, [user.uid]: score }
-          transaction.update(projectsRef.doc(id), { grades })
-        })
-      } catch (e) {
-        setShowError(true)
-        console.err(e)
-      }
+      await submitGrade(id, score, user.uid, setIsSubmitting, setShowError)
       setIsSubmitting(false)
       setSuccess(true)
       setTimeout(() => setLocation('/judging'), REDIRECT_TIMEOUT)
