+++ src/utility/firebase.js
@@ -139,6 +139,32 @@ export const getSubmission = async uid => {
   return (await projectsRef.doc(uid).get()).data()
 }
 
+export const submitGrade = async (
+  id,
+  score,
+  userUid,
+  submittingCallback = () => {},
+  errorCallback = () => {}
+) => {
+  try {
+    await db.runTransaction(async transaction => {
+      const projectDoc = await transaction.get(projectsRef.doc(id))
+      if (!projectDoc.exists) {
+        submittingCallback(false)
+        console.err('Project does not exist')
+        errorCallback(true)
+        return
+      }
+      const oldGrades = projectDoc.data().grades
+      const grades = { ...oldGrades, [userUid]: score }
+      transaction.update(projectsRef.doc(id), { grades })
+    })
+  } catch (e) {
+    errorCallback(true)
+    console.err(e)
+  }
+}
+
 export const updateUserApplication = async (uuid, newApp) => {
   return applicantsRef.doc(uuid).set(newApp)
 }
